feature(boiler-plate): updated doc
