+++ README.md
@@ -267,17 +267,11 @@ Please create `.env` file in root directory of the project (boilerplate-for-nft/
 The method of obtaining each value will now be explained.
 
 ```
-REACT_APP_DAPP_ID=<your dapp id>
-REACT_APP_API_KEY=<your api key>
+REACT_APP_CONTRACT_ADDRESS=<your contract address>
 ```
 
-DApp ID and API Key can be viewed by clicking Client SDK from the sidebar of the app.
+If you are using something other than Mumbai Testnet, change the following constants as needed in file `src/const/networkInfo`.
 
-<img width="238" alt="スクリーンショット 2022-10-11 14 46 08" src="https://user-images.githubusercontent.com/53442928/195006480-96fa2eda-f4ec-40e1-a631-a9d4ff19cd78.png">
-
-If you are using something other than Mumbai Testnet, change the following constants as needed.
-
-<img width="561" alt="スクリーンショット 2022-10-11 15 46 09" src="https://user-images.githubusercontent.com/53442928/195015633-da340e95-7049-41f2-bc47-4eb9011ab8dc.png">
 
 ### Start application
