+++ src/const/networkInfo.ts
@@ -1,0 +1,8 @@
+// ===== Make changes as needed (start) =====
+export const chainId = 80001;
+export const NETWORK_INFO = {
+  chainName: 'Mumbai Testnet',
+  chainId: '0x' + chainId.toString(16),
+  nativeCurrency: { name: 'MATIC', decimals: 18, symbol: 'MATIC' },
+  rpcUrls: ['https://rpc-mumbai.maticvigil.com'],
+};
