--- src/libs/fetcher.ts
@@ -1,7 +1,0 @@
-export const fetcher =
-  (library: any) =>
-  (...args: any[]) => {
-    const [method, ...params] = args;
-
-    return library[method](...params);
-  };
